requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.21.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
loguru==0.7.2
greenlet==3.2.3